fix  #1 («) operator

+++ ex01/main.cpp
@@ -6,7 +6,7 @@
 /*   By: tokazaki <tokazaki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/30 21:46:14 by tokazaki          #+#    #+#             */
-/*   Updated: 2023/12/13 14:33:26 by tokazaki         ###   ########.fr       */
+/*   Updated: 2023/12/16 14:39:23 by tokazaki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,14 +26,14 @@ int main() {
     std::cout << RED;
     a = Fixed(1234.4321f);
 
-    std::cout << RED << "a is " << a.toFloat() << std::endl;
-    std::cout << BLUE << "b is " << b.toInt() << std::endl;
-    std::cout << GREEN << "c is " << c.toFloat() << std::endl;
-    std::cout << CYAN << "d is " << d.toInt() << std::endl;
+    std::cout << RED << "a is " << a << std::endl;
+    std::cout << BLUE << "b is " << b << std::endl;
+    std::cout << GREEN << "c is " << c << std::endl;
+    std::cout << CYAN << "d is " << d << std::endl;
 
     std::cout << RED << "a is " << a.toInt() << " as integer" << std::endl;
     std::cout << BLUE << "b is " << b.toInt() << " as integer" << std::endl;
-    std::cout << GREEN <<"c is " << static_cast<int>(c.toFloat()) << " as integer" << std::endl;
+    std::cout << GREEN <<"c is " << c.toInt() << " as integer" << std::endl;
     std::cout << CYAN <<"d is " << d.toInt() << " as integer" << std::endl;
     std::cout << END;
